P2: normalize volatility to a rate instead of a capped count
windowed_volatility() counted skill-level changes among the last 5
readings and clamped the result at 3. A count cannot distinguish a
player who moved twice across three readings -- unsettled on every
opportunity he had -- from one who moved twice across five, who held
steady more often than not. Both reported a flat 2 and scored
identically.

normalized_volatility() divides by the opportunities to change:

    volatility        = changes / valid_transitions
    valid_transitions = window_length - 1

where window_length is how many readings are actually present, not the
nominal window size. Those two players now read 1.0 and 0.5.

Consequences, all of them deliberate:

- The cap is gone. A rate is bounded at 1.0 by construction, so
  clamping is no longer meaningful.
- Readings with no skill level are excluded from BOTH sides of the
  ratio. skill_level_changes already skipped them; leaving them in the
  denominator would have diluted a real 1.0 into 0.5.
- volatility_penalty() now scales a [0,1] rate across the same 0-15
  range it always used, rather than charging 3 per change, and clamps
  its input so a stale raw count cannot exceed the documented ceiling.
- confidence_score()'s volatility component is 100 * (1 - rate), which
  preserves its 0-100 range.
- Matchup.volatility becomes Float. As Integer it would have silently
  truncated every partial rate to 0 -- the entire point of the change,
  lost at the database boundary.

skill_level_volatility() is untouched: it still backs the unrelated
Skill Level History summary as a whole-history, uncapped count, and
that separation was deliberate before this change and remains so.

Tests cover both required properties directly: the same raw change
count over different history lengths yields different rates, and the
same rate yields identical matchup_score, confidence_score and
penalty -- with a companion assertion that a DIFFERENT rate still
scores differently, so the equivalence cannot pass by the score having
stopped responding to volatility at all.

Scope: P2 only. No Gate A files, no VBA or template work.

392 tests passing.

+++ database/models.py
@@ -377,7 +377,9 @@ class PlayerMatchup(Base):
     # looked overall", not "did a skill-gap win earn a bonus".
     sl_delta = Column(Float)
     trend = Column(String)
-    volatility = Column(Integer)
+    # P2: a normalized rate in [0.0, 1.0], not a change count -- Integer
+    # here would silently truncate every partial rate to 0.
+    volatility = Column(Float)
     matchup_score = Column(Integer)
     confidence_score = Column(Integer)
     format = Column(String)
